gfdg

+++ DiceApi.Data/Data/User/Api/Model/UserApi.cs
@@ -16,7 +16,7 @@ namespace DiceApi.Data.Api.Model
         public string Name { get; set; }
 
         [JsonProperty("ballance")]
-        public double Ballance { get; set; }
+        public decimal Ballance { get; set; }
 
         [JsonProperty("ownerId")]
         public long? OwnerId { get; set; }
